+++ apps/next-client/src/app/community/[id]/edit/page.tsx
@@ -1,13 +1,9 @@
 'use client';
 
-import React, { useEffect, useState, useMemo } from 'react';
-import { useRouter } from 'next/navigation';
-import { axiosInstance } from '@/services/axiosInstance';
+import React, { useMemo } from 'react';
 import dynamic from 'next/dynamic';
 import { useSession } from 'next-auth/react';
-import { API_URLS } from '@/constants/urls';
-import { ALERT_MESSAGES } from '@/constants/alertMessage';
-import { useQueryClient } from '@tanstack/react-query';
+import { usePost } from '@/hooks/queries/usePostEdit';
 import '@/styles/pages/community/community.scss';
 
 const ReactQuill = dynamic(() => import('react-quill'), { ssr: false });
@@ -15,42 +11,12 @@ import 'react-quill/dist/quill.snow.css';
 
 export default function EditPostPage({ params }: { params: { id: string } }) {
   const { id } = params;
-  const [title, setTitle] = useState('');
-  const [content, setContent] = useState('');
-  const [loading, setLoading] = useState(true);
-
   const { data: session } = useSession();
-  const router = useRouter();
-  const queryClient = useQueryClient();
 
   const userId = session?.user?.id;
-  const accessToken = session?.user?.accessToken;
+  const accessToken = session?.user?.accessToken || '';
 
-  useEffect(() => {
-    const fetchPost = async () => {
-      try {
-        const response = await axiosInstance.get(`${API_URLS.POSTS}/${id}`);
-        const post = response.data;
-
-        if (post.userId !== userId) {
-          alert(ALERT_MESSAGES.ERROR.POST.POST_PERMISSION_DENIED);
-          router.push('/community');
-          return;
-        }
-
-        setTitle(post.title);
-        setContent(post.content);
-      } catch (error) {
-        console.error('Error fetching post:', error);
-        alert(ALERT_MESSAGES.ERROR.POST.POST_FETCH_ERROR);
-        router.push('/community');
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchPost();
-  }, [id, userId, router]);
+  const { title, setTitle, content, setContent, loading, handleUpdatePost, handleDeletePost } = usePost(id, userId, accessToken);
 
   const modules = useMemo(() => {
     return {
@@ -88,46 +54,6 @@ export default function EditPostPage({ params }: { params: { id: string } }) {
     ],
     []
   );
-
-  const handleUpdatePost = async () => {
-    try {
-      if (!title.trim() || !content.trim()) {
-        alert(ALERT_MESSAGES.ERROR.POST.POST_EMPTY_FIELDS);
-        return;
-      }
-
-      await axiosInstance.put(
-        `${API_URLS.POSTS}/${id}`,
-        { title, content },
-        { headers: { Authorization: `Bearer ${accessToken}` } }
-      );
-
-      alert(ALERT_MESSAGES.SUCCESS.POST.POST_UPDATE);
-
-      queryClient.invalidateQueries({ queryKey: ['post', id] });
-
-      router.push(`/community/${id}`);
-    } catch (error) {
-      console.error('Error updating post:', error);
-      alert(ALERT_MESSAGES.ERROR.POST.POST_UPDATE_ERROR);
-    }
-  };
-
-  const handleDeletePost = async () => {
-    try {
-      if (!window.confirm(ALERT_MESSAGES.CONFIRM.CHECK_DELETE)) return;
-
-      await axiosInstance.delete(`${API_URLS.POSTS}/${id}`, {
-        headers: { Authorization: `Bearer ${accessToken}` },
-      });
-
-      alert(ALERT_MESSAGES.SUCCESS.POST.POST_DELETE);
-      router.push('/community');
-    } catch (error) {
-      console.error('Error deleting post:', error);
-      alert(ALERT_MESSAGES.ERROR.POST.POST_DELETE_ERROR);
-    }
-  };
 
   if (loading) {
     return <p>게시글을 불러오는 중...</p>;
